requirements.txt
numpy==1.26.4
opencv_python==4.8.1.78
pandas==2.2.3
Pillow==10.1.0
plotly==5.23.0
python_dateutil==2.9.0.post0
scikit_learn==1.3.2
scipy==1.15.1
streamlit==1.37.1
streamlit_image_comparison==0.0.4
torch==2.4.0
tqdm==4.66.5
